pyproject.toml
[tool.poetry]
name = "rag-agentic-chatbot"
version = "0.1.0"
description = "A hallucination-proof RAG chatbot using LangGraph"
authors = ["Antigravity"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.1.0"
langgraph = "^0.0.10"
langchain-google-genai = "^0.0.5"
langchain-community = "^0.0.10"
chromadb = ">=0.5.0"
pandas = "^2.2.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
